pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "myjob"
version = "0.1.0"
description = "CLI tool for submitting and managing SLURM jobs on remote clusters"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "jhbyun" }
]
dependencies = [
    "typer>=0.9.0",
    "fabric>=3.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
myjob = "myjob.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["myjob"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.10"
strict = true
